+++ src/python_example/depth_estimation/yolo26_depth_s/factory/__init__.py
@@ -1,0 +1,3 @@
+from .yolo26_depth_s_factory import Yolo26DepthSFactory
+
+__all__ = ["Yolo26DepthSFactory"]